+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Fixes for energy settings
+
 ## 1.3.2 (2022-11-05)
 
 * (theimo1221) Fix for sceneSettings Persistence